requirements.txt
astroid==2.11.2
dataclasses==0.6
dill==0.3.4
distlib==0.3.4
filelock==3.6.0
flake8==4.0.1
gitdb==4.0.9
GitPython==3.1.27
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
mypy==0.942
mypy-extensions==0.4.3
packaging==21.3
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pylint==2.13.4
pyparsing==3.0.7
six==1.16.0
smmap==5.0.0
toml==0.10.2
tomli==2.0.1
tox==3.24.5
tqdm==4.64.0
typing_extensions==4.1.1
virtualenv==20.14.0
wrapt==1.14.0
